+++ openghg_inversions/postprocessing/countries.py
@@ -16,7 +16,7 @@ from openghg_inversions import convert, utils
 from openghg_inversions.array_ops import align_sparse_lat_lon, get_xr_dummies, sparse_xr_dot
 from openghg_inversions.utils import get_country_file_path
 from ._country_codes import CountryInfoList
-from .inversion_output import InversionOutput
+from .inversion_output import LegacyInversionOutput
 
 # type for xr.Dataset *or* xr.DataArray
 DataSetOrArray = TypeVar("DataSetOrArray", xr.DataArray, xr.Dataset)
@@ -366,13 +366,13 @@ class Countries:
 
     def get_x_to_country_mat(
         self,
-        inv_out: InversionOutput,
+        inv_out: LegacyInversionOutput,
         sparse: bool = False,
     ) -> xr.DataArray:
         """Construct a sparse matrix mapping from x sensitivities to country totals.
 
         Args:
-            inv_out: InversionOutput object, used to get basis functions and flux.
+            inv_out: LegacyInversionOutput object, used to get basis functions and flux.
             sparse: if True, values of returned DataArray are `sparse.COO` array.
 
         Returns:
@@ -420,13 +420,13 @@ class Countries:
 
     def get_country_trace(
         self,
-        inv_out: InversionOutput,
+        inv_out: LegacyInversionOutput,
     ) -> xr.Dataset:
         """Calculate trace(s) for total country emissions.
 
         Args:
             species: name of species, e.g. "co2", "ch4", "sf6", etc.
-            inv_out: InversionOutput
+            inv_out: LegacyInversionOutput
 
         Returns:
             xr.Dataset with coordinate dimensions ("country", "draw")